# ShadingEngine : Prototype support for OSL SIMD