Putting up a draft of where we're at with OSL SIMD support.  This now appears to be working, as long as you've got OSL 1.12.8.0 + this PR https://github.com/AcademySoftwareFoundation/OpenShadingLanguage/pull/1637

Main outstanding issues:
* Should I add ifdefs so that this can be merged immediately, and at least compile without OSL 1.12 ( it can't work without it, but currently it won't even compile )
* How should this be enabled?  Checkbox on OSLImage and OSLObject?  Default on or off?
* How should you check a diagnostic for what style of batched execution you are actually getting if you do turn it on?  If anyone has questions about performance, it's important to know whether they are actually getting batches of 8 or 16, and which CPU arch - getting the right CPU arch depends on having built OSL with the right build flags, so it's very possible for someone to not be getting any speedup, or only be getting partial speedup, and it seems important to be able to diagnose this.