@danieldresser-ie 2023-01-19
As for performance, I haven't tested in too much detail, since we don't really control the performance of this - the goal is just to turn it on, and benefit from whatever speedup OSL gives.  But I've been doing basic testing with this artificially easy to optimize test:

```
import Gaffer
import GafferImage
import GafferOSL
import IECore
import imath

__children = {}

__children["OSLImage"] = GafferOSL.OSLImage( "OSLImage" )
parent.addChild( __children["OSLImage"] )
__children["OSLImage"]["channels"].addChild( Gaffer.NameValuePlug( "", Gaffer.Color3fPlug( "value", defaultValue = imath.Color3f( 1, 1, 1 ), flags = Gaffer.Plug.Flags.Default | Gaffer.Plug.Flags.Dynamic, ), True, "channel", Gaffer.Plug.Flags.Default | Gaffer.Plug.Flags.Dynamic ) )
__children["OSLCode"] = GafferOSL.OSLCode( "OSLCode" )
parent.addChild( __children["OSLCode"] )
__children["OSLCode"]["parameters"].addChild( Gaffer.FloatPlug( "scale", defaultValue = 0.0, flags = Gaffer.Plug.Flags.Default | Gaffer.Plug.Flags.Dynamic, ) )
__children["OSLCode"]["parameters"].addChild( Gaffer.V3fPlug( "offset", defaultValue = imath.V3f( 0, 0, 0 ), flags = Gaffer.Plug.Flags.Default | Gaffer.Plug.Flags.Dynamic, interpretation = IECore.GeometricData.Interpretation.Vector ) )
__children["OSLCode"]["out"].addChild( Gaffer.Color3fPlug( "out", direction = Gaffer.Plug.Direction.Out, defaultValue = imath.Color3f( 0, 0, 0 ), flags = Gaffer.Plug.Flags.Default | Gaffer.Plug.Flags.Dynamic, ) )
__children["OSLImage"]["defaultFormat"]["displayWindow"]["min"].setValue( imath.V2i( 0, 0 ) )
__children["OSLImage"]["defaultFormat"]["displayWindow"]["max"]["x"].setValue( 4096 )
Gaffer.Metadata.registerValue( __children["OSLImage"]["defaultFormat"], 'formatPlugValueWidget:mode', 'custom' )
__children["OSLImage"]["defaultFormat"]["displayWindow"]["max"]["y"].setInput( __children["OSLImage"]["defaultFormat"]["displayWindow"]["max"]["x"] )
__children["OSLImage"]["channels"]["channel"]["value"].setInput( __children["OSLCode"]["out"]["out"] )
__children["OSLCode"]["parameters"]["scale"].setValue( -2578.32958984375 )
__children["OSLCode"]["parameters"]["offset"].setValue( imath.V3f( -0.379400015, -0.0984999985, 0 ) )
__children["OSLCode"]["code"].setValue( 'vector p = color( u, v, 0 ) - vector( 0.5 );\np = p / scale - offset;\nvector a = 0;\nfor( int i = 0; i < 200; i++ )\n{\n\ta = vector( a[0] * a[0] - a[1] * a[1] + p[0], 2 * a[0] * a[1] + p[1], 0 );\n}\nout = a;' )

del __children
```

Running that in Gaffer stats yields:
Without SIMD:
  Image generation    0.574s (wall), 8.097s (CPU)
With SIMD:
  Image generation    1.172s (wall), 18.099s (CPU)  

About a 2X speedup.  Noting that the CPU times were showing even a bit more difference suggested we were seeing some overhead here - turning up the mandelbrot iterations to 800 showed a 3X speedup.

I wouldn't expect gains this significant in most reasonable situations ( especially not before we hook up arbitrary pixel sampling, which will make it easier to find very expensive things to do ), but this does seem to demonstrate that the SIMD support is working, and can make a significant difference.