@johnhaddon 2023-01-19
Nice one! At this point I've only skimmed the changes to give me an idea of how things look, to help answer your questions 

> Should I add ifdefs so that this can be merged immediately, and at least compile without OSL 1.12

That seems like a good idea to me - better than keeping the code in limbo. If we think we're close on this, and OSL will release a version with the appropriate fixes soon, then it seems like we should aim to get this in Gaffer 1.2.0.0. @ericmehl, do you foresee the required OSL upgrade causing any problems for the Windows release? @ivanimanishi, do you think an OSL update could be on the cards for the IE builds of 1.2.0.0, or will you definitely require source compatibility for a non-simd OSL version?

> How should this be enabled? Checkbox on OSLImage and OSLObject?

My assumption is that once proven stable, SIMD should always be on, and single-point execution should be a thing of the past (and we should rip out the code for it). That makes me think that an environment variable like our `GAFFER_HASHCACHE_MODE` is the most suitable configuration mechanism. It would default to the new goodness, giving the best performance we can, but folks could turn it off if there turned out to be problems adopting it in production.

I can see that a plug would also be useful for interactive testing and development though. But I don't think we'd really want users to be aware of it, because we'd want to remove it in the future without breaking old files. So perhaps a non-serialisable plug that defaults to hidden in the UI?

> Default on or off?

Default on sounds right to me, so we get maximum performance benefit and maximum testing. But we should do some testing with production setups before release to give us some confidence in that decision.

> How should you check a diagnostic for what style of batched execution you are actually getting if you do turn it on? If anyone has questions about performance, it's important to know whether they are actually getting batches of 8 or 16, and which CPU arch - getting the right CPU arch depends on having built OSL with the right build flags, so it's very possible for someone to not be getting any speedup, or only be getting partial speedup, and it seems important to be able to diagnose this.

Throw an exception if SIMD was requested but isn't available? Plus a DEBUG level message with the full details of width and arch?