@ivanimanishi 2023-01-19
> @ivanimanishi, do you think an OSL update could be on the cards for the IE builds of 1.2.0.0, or will you definitely require source compatibility for a non-simd OSL version?

After a chat with Daniel, I think that we are likely to be able to use osl 1.12. But I'd be much more at ease if we did support 1.11 just in case.