@ericmehl 2023-01-19
> do you foresee the required OSL upgrade causing any problems for the Windows release

It looks like it should be pretty smooth to update on Windows. We're already meeting or exceeding the new minimum versions for dependencies and the compiler.