@boberfly 2023-01-19
+1 from me, v1.12.8.0 comes with some important patches for cycles OSL GPU support on optix for when we eventually enable that one.