@danieldresser-ie 2023-01-20
Took a first swing at implementing your suggestions.  It's now on by default, but disabled by setting GAFFER_OSL_USESIMD=0.  I haven't implemented a plug yet ... it could be handy for testing, but if we don't want to expose it to users anyway, it's a little annoying to implement ... the env var works.

I've put the info about what arch you're actually getting in an INFO level message - it seems more valuable than some of the extreme minutiae that ends up in DEBUG level?

I've added backwards compatibility with old OSL - I'm not sure if it's a weird final situation where when combined with your other suggestion, if you use this with the version of OSL currently in gafferDependencies, it will error because it's unable to use SIMD, until you set the env var to 0?