@danieldresser-ie 2023-01-27
OK, I think I've addressed all specific notes.  Let me know how it looks, and I can at least squash it.  In terms of next steps, I guess we need to wait for the next OSL release anyway ... and I still feel like I should be testing it more given how much weirdness I've found so far, but maybe we just have to put out a testing release and see how it goes?