@danieldresser-ie 2023-01-31
As per request from John, squashed everything and added a Changelog entry so this is hypothetically ready for merge.